Remove annoying "peculiar error" message when requests are aborted
`request' often spams the *Messages* buffer with this useless message:

    [error] request--callback: peculiar error

This happens when a live request is interrupted/aborted.  This change
disables logging of that message for CE requests, as we often have to
abort the requests.

* compiler-explorer.el (compiler-explorer--requests): New var.
(compiler-explorer--request): Use it.
(compiler-explorer--request-callback-wrapper): New function.
(compiler-explorer-new-session): Use it as an :around wrapper for
`request-callback'.